requirements.txt
--index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments
psycopg2-binary
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.10.1
yfinance
multitasking==0.0.11
pendulum
pandas